Add condition to DELETE user

+++ index.js
@@ -159,6 +159,11 @@ app.delete('/users/:Username/movies/:MovieID', passport.authenticate('jwt', { se
 
 // DELETE user by Username
 app.delete('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    // Condition to check user authorization
+    if(req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }
+    // Condition ends here
     await Users.findOneAndDelete({ Username: req.params.Username })
         .then((user) => {
             if (!user) {
